[SYCL][UR][L0] Add UR_L0 environment variables
Define UR_LO_DEBUG and UR_L0_SERIALIZE, and keep
their ZE_ counterparts for now.

And do some reorganization
