+++ sycl/plugins/unified_runtime/ur/adapters/level_zero/ur_level_zero_common.cpp
@@ -59,3 +59,214 @@ ur_result_t ze2urResult(ze_result_t ZeResult) {
     return UR_RESULT_ERROR_UNKNOWN;
   }
 }
+
+void urPrint(const char *Format, ...) {
+  if (UrL0Debug & UR_L0_DEBUG_BASIC) {
+    va_list Args;
+    va_start(Args, Format);
+    vfprintf(stderr, Format, Args);
+    va_end(Args);
+  }
+}
+
+void zePrint(const char *Format, ...) {
+  if (ZeDebug & ZE_DEBUG_BASIC) {
+    va_list Args;
+    va_start(Args, Format);
+    vfprintf(stderr, Format, Args);
+    va_end(Args);
+  }
+}
+
+// This function will ensure compatibility with both Linux and Windows for
+// setting environment variables.
+bool setEnvVar(const char *name, const char *value) {
+#ifdef _WIN32
+  int Res = _putenv_s(name, value);
+#else
+  int Res = setenv(name, value, 1);
+#endif
+  if (Res != 0) {
+    zePrint(
+        "Level Zero plugin was unable to set the environment variable: %s\n",
+        name);
+    return false;
+  }
+  return true;
+}
+
+// This will count the calls to Level-Zero
+std::map<const char *, int> *ZeCallCount = nullptr;
+
+inline void zeParseError(ze_result_t ZeError, const char *&ErrorString) {
+  switch (ZeError) {
+#define ZE_ERRCASE(ERR)                                                        \
+  case ERR:                                                                    \
+    ErrorString = "" #ERR;                                                     \
+    break;
+
+    ZE_ERRCASE(ZE_RESULT_SUCCESS)
+    ZE_ERRCASE(ZE_RESULT_NOT_READY)
+    ZE_ERRCASE(ZE_RESULT_ERROR_DEVICE_LOST)
+    ZE_ERRCASE(ZE_RESULT_ERROR_OUT_OF_HOST_MEMORY)
+    ZE_ERRCASE(ZE_RESULT_ERROR_OUT_OF_DEVICE_MEMORY)
+    ZE_ERRCASE(ZE_RESULT_ERROR_MODULE_BUILD_FAILURE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INSUFFICIENT_PERMISSIONS)
+    ZE_ERRCASE(ZE_RESULT_ERROR_NOT_AVAILABLE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_UNINITIALIZED)
+    ZE_ERRCASE(ZE_RESULT_ERROR_UNSUPPORTED_VERSION)
+    ZE_ERRCASE(ZE_RESULT_ERROR_UNSUPPORTED_FEATURE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_ARGUMENT)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_NULL_HANDLE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_HANDLE_OBJECT_IN_USE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_NULL_POINTER)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_SIZE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_UNSUPPORTED_SIZE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_UNSUPPORTED_ALIGNMENT)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_SYNCHRONIZATION_OBJECT)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_ENUMERATION)
+    ZE_ERRCASE(ZE_RESULT_ERROR_UNSUPPORTED_ENUMERATION)
+    ZE_ERRCASE(ZE_RESULT_ERROR_UNSUPPORTED_IMAGE_FORMAT)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_NATIVE_BINARY)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_GLOBAL_NAME)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_KERNEL_NAME)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_FUNCTION_NAME)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_GROUP_SIZE_DIMENSION)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_GLOBAL_WIDTH_DIMENSION)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_KERNEL_ARGUMENT_INDEX)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_KERNEL_ARGUMENT_SIZE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_KERNEL_ATTRIBUTE_VALUE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_COMMAND_LIST_TYPE)
+    ZE_ERRCASE(ZE_RESULT_ERROR_OVERLAPPING_REGIONS)
+    ZE_ERRCASE(ZE_RESULT_ERROR_INVALID_MODULE_UNLINKED)
+    ZE_ERRCASE(ZE_RESULT_ERROR_UNKNOWN)
+
+#undef ZE_ERRCASE
+  default:
+    assert(false && "Unexpected Error code");
+  } // switch
+}
+
+ze_result_t ZeCall::doCall(ze_result_t ZeResult, const char *ZeName,
+                           const char *ZeArgs, bool TraceError) {
+  zePrint("ZE ---> %s%s\n", ZeName, ZeArgs);
+
+  if (ZeDebug & ZE_DEBUG_CALL_COUNT) {
+    ++(*ZeCallCount)[ZeName];
+  }
+
+  if (ZeResult && TraceError) {
+    const char *ErrorString = "Unknown";
+    zeParseError(ZeResult, ErrorString);
+    zePrint("Error (%s) in %s\n", ErrorString, ZeName);
+  }
+  return ZeResult;
+}
+
+// Specializations for various L0 structures
+template <> ze_structure_type_t getZeStructureType<ze_event_pool_desc_t>() {
+  return ZE_STRUCTURE_TYPE_EVENT_POOL_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_fence_desc_t>() {
+  return ZE_STRUCTURE_TYPE_FENCE_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_command_list_desc_t>() {
+  return ZE_STRUCTURE_TYPE_COMMAND_LIST_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_context_desc_t>() {
+  return ZE_STRUCTURE_TYPE_CONTEXT_DESC;
+}
+template <>
+ze_structure_type_t
+getZeStructureType<ze_relaxed_allocation_limits_exp_desc_t>() {
+  return ZE_STRUCTURE_TYPE_RELAXED_ALLOCATION_LIMITS_EXP_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_host_mem_alloc_desc_t>() {
+  return ZE_STRUCTURE_TYPE_HOST_MEM_ALLOC_DESC;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_device_mem_alloc_desc_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_MEM_ALLOC_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_command_queue_desc_t>() {
+  return ZE_STRUCTURE_TYPE_COMMAND_QUEUE_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_image_desc_t>() {
+  return ZE_STRUCTURE_TYPE_IMAGE_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_module_desc_t>() {
+  return ZE_STRUCTURE_TYPE_MODULE_DESC;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_module_program_exp_desc_t>() {
+  return ZE_STRUCTURE_TYPE_MODULE_PROGRAM_EXP_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_kernel_desc_t>() {
+  return ZE_STRUCTURE_TYPE_KERNEL_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_event_desc_t>() {
+  return ZE_STRUCTURE_TYPE_EVENT_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_sampler_desc_t>() {
+  return ZE_STRUCTURE_TYPE_SAMPLER_DESC;
+}
+template <> ze_structure_type_t getZeStructureType<ze_driver_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DRIVER_PROPERTIES;
+}
+template <> ze_structure_type_t getZeStructureType<ze_device_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_device_compute_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_COMPUTE_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_command_queue_group_properties_t>() {
+  return ZE_STRUCTURE_TYPE_COMMAND_QUEUE_GROUP_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_device_image_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_IMAGE_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_device_module_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_MODULE_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_device_cache_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_CACHE_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_device_memory_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_MEMORY_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_device_memory_ext_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_MEMORY_EXT_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_device_memory_access_properties_t>() {
+  return ZE_STRUCTURE_TYPE_DEVICE_MEMORY_ACCESS_PROPERTIES;
+}
+template <> ze_structure_type_t getZeStructureType<ze_module_properties_t>() {
+  return ZE_STRUCTURE_TYPE_MODULE_PROPERTIES;
+}
+template <> ze_structure_type_t getZeStructureType<ze_kernel_properties_t>() {
+  return ZE_STRUCTURE_TYPE_KERNEL_PROPERTIES;
+}
+template <>
+ze_structure_type_t getZeStructureType<ze_memory_allocation_properties_t>() {
+  return ZE_STRUCTURE_TYPE_MEMORY_ALLOCATION_PROPERTIES;
+}
+
+template <> zes_structure_type_t getZesStructureType<zes_pci_properties_t>() {
+  return ZES_STRUCTURE_TYPE_PCI_PROPERTIES;
+}
+
+template <> zes_structure_type_t getZesStructureType<zes_mem_state_t>() {
+  return ZES_STRUCTURE_TYPE_MEM_STATE;
+}
+
+template <> zes_structure_type_t getZesStructureType<zes_mem_properties_t>() {
+  return ZES_STRUCTURE_TYPE_MEM_PROPERTIES;
+}
